+++ shared/flow.py
@@ -85,10 +85,8 @@ def se2_and_real_secret():
     from pincodes import pa
     return (not pa.is_secret_blank()) and (not pa.tmp_value)
 
-def bip39_passphrase_active():
-    import stash
-    return settings.get('words', True) \
-        or (settings.master_get('words', True) and stash.bip39_passphrase)
+def word_based_seed():
+    return settings.get("words", True)
 
 
 HWTogglesMenu = [
@@ -243,7 +241,7 @@ DebugFunctionsMenu = [
 
 SeedXORMenu = [
     #         xxxxxxxxxxxxxxxx
-    MenuItem("Split Existing", f=xor_split_start, predicate=lambda: settings.get('words', True)),
+    MenuItem("Split Existing", f=xor_split_start, predicate=word_based_seed),
     MenuItem("Restore Seed XOR", f=xor_restore_start),
 ]
 
@@ -252,8 +250,7 @@ SeedFunctionsMenu = [
     MenuItem('Seed XOR', menu=SeedXORMenu),
     MenuItem("Destroy Seed", f=clear_seed, predicate=se2_and_real_secret),
     MenuItem('Lock Down Seed', f=convert_ephemeral_to_master),
-    MenuItem('Export SeedQR', f=export_seedqr,
-             predicate=lambda: settings.get('words', True)),
+    MenuItem('Export SeedQR', f=export_seedqr, predicate=word_based_seed),
 ]
 
 DangerZoneMenu = [
@@ -382,7 +379,7 @@ EmptyWallet = [
 NormalSystem = [
     #         xxxxxxxxxxxxxxxx
     MenuItem('Ready To Sign', f=ready2sign, shortcut='r'),
-    MenuItem('Passphrase', f=start_b39_pw, predicate=bip39_passphrase_active, shortcut='p'),
+    MenuItem('Passphrase', f=start_b39_pw, predicate=word_based_seed, shortcut='p'),
     MenuItem('Scan Any QR Code', predicate=lambda: version.has_qr,
          shortcut=KEY_QR, f=scan_any_qr, arg=(False, True)),
     MenuItem('Start HSM Mode', f=start_hsm_menu_item, predicate=hsm_policy_available),